:art: onBtnCarNamesClicked() 모듈로 분리
