+++ src/library/car-names.js
@@ -1,3 +1,6 @@
+import { makeElementBlock } from "./make-element-block.js";
+import { Car } from "./car.js";
+
 export function isCorrectCarNames(carName) {
   if (carName.length > 0 && carName.length <= 5) {
     return true;
@@ -6,7 +9,7 @@ export function isCorrectCarNames(carName) {
   return false;
 }
 
-export function addCorrectCarNames(carNames) {
+export function addCorrectCarNames(carNames, inputCarNames) {
   const correctCarNames = carNames.map(function (carName) {
     carName = carName.trim();
     if (!isCorrectCarNames(carName)) {
@@ -20,4 +23,19 @@ export function addCorrectCarNames(carNames) {
   });
 
   return correctCarNames;
+}
+
+export function onBtnCarNamesClicked(cars, inputCarNames) {
+  const carNames = inputCarNames.value.split(",");
+  const correctCarNames = addCorrectCarNames(carNames, inputCarNames);
+  const racingBox = document.getElementById("racing-count-box");
+
+  if (!correctCarNames.includes(undefined)) {
+    makeElementBlock(racingBox);
+    correctCarNames.forEach((correctCar) => {
+      cars.push(new Car(correctCar));
+    });
+  }
+
+  return;
 }
